[NumPy] Remove references to deprecated NumPy type aliases (#21022)
[NumPy] Remove references to deprecated NumPy type aliases.

This change replaces references to a number of deprecated NumPy type aliases (np.bool, np.int, np.float, np.complex, np.object, np.str) with their recommended replacement (bool, int, float, complex, object, str).

NumPy 1.24 drops the deprecated aliases, so we must remove uses before updating NumPy.

+++ examples/tensorflow/benchmarking/plot_csv_file.py
@@ -132,7 +132,7 @@ class Plot:
                 if self.args.plot_along_batch:
                     y_axis_array = np.asarray(
                         [results[(x, inner_loop_value)] for x in x_axis_array if (x, inner_loop_value) in results],
-                        dtype=np.int,
+                        dtype=int,
                     )
                 else:
                     y_axis_array = np.asarray(
@@ -144,7 +144,7 @@ class Plot:
                     ("batch_size", "len") if self.args.plot_along_batch else ("in #tokens", "bsz")
                 )
 
-                x_axis_array = np.asarray(x_axis_array, np.int)[: len(y_axis_array)]
+                x_axis_array = np.asarray(x_axis_array, int)[: len(y_axis_array)]
                 plt.scatter(
                     x_axis_array, y_axis_array, label=f"{label_model_name} - {inner_loop_label}: {inner_loop_value}"
                 )
